consumer-service/requirements.txt
pika==1.3.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1